+++ user-management-app/uma-server/server.js
@@ -1,11 +1,21 @@
-const express = require('express');
+// server.js
+const express = require("express");
+const bodyParser = require("body-parser");
+const customCorsMiddleware = require("./middleware/customCorsMiddleware");
+const userRouter = require("./routes/userRouter"); // Import the user router
+require("dotenv").config();
+
 const app = express();
-const port = 3000;
+const NODE_APP_PORT = process.env.SERVER_PORT;
 
-app.get('/', (req, res) => {
-  res.send('Hello World!');
-});
+// Middleware
+app.use(bodyParser.json());
+app.use(customCorsMiddleware);
 
-app.listen(port, () => {
-  console.log(`Server listening at http://localhost:${port}`);
-});
+// Use the user router for user-related routes
+app.use("/api/user", userRouter);
+
+// Start server
+app.listen(NODE_APP_PORT, () =>
+  console.log(`Node server running on ${NODE_APP_PORT}`)
+);