+++ user-management-app/CHANGELOG.md
@@ -4,8 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.0.1] - 2024-02-20
+
+### Added some dependencies
+- npm install mssql 
+- npm install body-parser
+- npm install dotenv
+- npm install cors
+- npm install -g nodemon
+
 ## [1.0.0] - 2024-02-20
 
 ### Added
 - Initial project setup
 - Created basic folder structure
+